Extract out query conversion utils
